requirements.txt
flask
GitPython
requests
PyGithub
Flask-SQLAlchemy
Flask-Caching
boto3
python-magic
flask-jwt-extended
cryptography
bcrypt
PyYAML
jsonschema
